Remove registry check for WindowsMediaVersion
Instead move the check for the MF startup result to CheckDependencies. It was used only for logging, but the logging was configured *after* the constructor call so it never got logged.

The check also probably never applied, because the loading of ScreenRecorderLib already fails if MF.dll is missing.

Fixes #227